requirements.txt
# Minimal requirements for training/prediction
torch
torchvision
numpy
pillow
